requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.0
python-dotenv==1.0.1
email-validator==2.1.1
Werkzeug==3.0.2
cryptography==42.0.5
google-generativeai==0.4.1
